agent/requirements.txt
# Orbit AI Desktop Agent - Dependencies
# ======================================

# Core
pynput>=1.7.6          # Keyboard & mouse capture (cross-platform)
mss>=9.0.1             # Fast screen capture
Pillow>=10.0.0         # Image processing
psutil>=5.9.0          # Process & system monitoring
watchdog>=3.0.0        # File system monitoring

# OCR
pytesseract>=0.3.10    # OCR (requires system tesseract)

# AI Analysis
anthropic>=0.39.0      # Claude Haiku API

# Storage
google-auth>=2.23.0    # Google OAuth2
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0  # Google Drive API

# Utilities
pyperclip>=1.8.2       # Clipboard access
apscheduler>=3.10.4    # Task scheduling
requests>=2.31.0       # HTTP client

# Windows-specific (installed conditionally)
# pywin32>=306         # Windows API access
# pygetwindow>=0.0.9   # Window management
